feat: auto-generate agent OG images on test submission (fixes #403)
Extract buildAgentSVG and PNG rendering into shared lib/og-gen.js.
Call generateAgentOG() asynchronously (via setImmediate) after saving
test results in both ABTI and SBTI POST handlers.

- New agents get personalized OG images immediately
- Non-blocking: errors logged, never crash the response
- og/agents/ directory created automatically if missing
- scripts/generate-agent-og.js now imports from lib/og-gen.js
